+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project follows [Semantic Versioning](http://semver.org/).
 ### Added
 - Display of minimum and maximum values that exist in the database
   for a date range criterion
+- npm task "build:stats" to run webpack-bundle-analyzer
+### Changed
+- Bundle size reduced by removing moment locales
 
 ## [1.12.0] - 2020-08-24
 ### Changed